Now has a black border around a white screen which a green blinking cursor.
Pressing E will drop a black block, but it disappears when one of WASD is
pressed. Which sucks.

I've started with the rearranging of existing code to make saving a block
possible, but I'm unsure as to how to read the state of an existing block.

What I tried, but didn't work, remains in comments. WARNING: DON'T
UNCOMMENT THESE WITHOUT FIXING THEM - IT GOES HORRIBLY, HORRIBLY WRONG! I
just left them as a reminder that it definitely doesn't work that way.